step5 - refactor domain package 로 이동
