+++ src/main/java/study/racing4/domain/Car.java
@@ -1,4 +1,4 @@
-package study.racing4;
+package study.racing4.domain;
 
 public class Car {
