+++ src/main/java/study/racing4/domain/CarCollection.java
@@ -1,4 +1,4 @@
-package study.racing4;
+package study.racing4.domain;
 
 import java.util.ArrayList;
 import java.util.List;
