+++ src/main/java/study/racing4/domain/CarRacing.java
@@ -1,4 +1,4 @@
-package study.racing4;
+package study.racing4.domain;
 
 import java.util.List;
