+++ src/main/java/study/racing4/domain/MovingStrategy.java
@@ -1,4 +1,4 @@
-package study.racing4;
+package study.racing4.domain;
 
 public interface MovingStrategy {
