+++ src/main/java/study/racing4/domain/RandomMovingStrategy.java
@@ -1,4 +1,4 @@
-package study.racing4;
+package study.racing4.domain;
 
 import java.util.Random;
