+++ src/test/java/study/racing4/RacingInputTest.java
@@ -2,6 +2,7 @@ package study.racing4;
 
 import org.junit.jupiter.api.DisplayName;
 import org.junit.jupiter.api.Test;
+import study.racing4.domain.RacingInput;
 
 import static org.assertj.core.api.Assertions.assertThat;
 import static org.assertj.core.api.Assertions.assertThatExceptionOfType;
